Simplified function structure, removed redundant dependencies
