+++ lib/actions/upsertPersonOrOrganization.js
@@ -13,10 +13,10 @@ See the License for the specific language governing permissions and
 limitations under the License.
  */
 
-const Q = require('q');
-const { newMessage } = require('../helpers');
 const { resolve } = require('./../utils/resolver');
-const { upsertObject, fetchUidFromReference, populateRelations } = require('./../utils/helpers');
+const {
+  upsertObject, fetchUidFromReference, populateRelations, newMessage,
+} = require('./../utils/helpers');
 const { getToken } = require('./../utils/authentication');
 
 /**
@@ -26,14 +26,13 @@ const { getToken } = require('./../utils/authentication');
  * @param {Object} cfg - configuration that is account information and configuration field values
  */
 async function processAction(msg, cfg) {
-  // const token = cfg.API_KEY;
-  const token = await getToken(cfg);
-  const self = this;
-  const oihUid = (msg.body.meta !== undefined && msg.body.meta.oihUid !== undefined) ? msg.body.meta.oihUid : 'oihUid not set yet';
-  let recordUid = (msg.body !== undefined && msg.body.meta.recordUid !== undefined) ? msg.body.meta.recordUid : undefined;
-  const applicationUid = (msg.body.meta !== undefined && msg.body.meta.applicationUid !== undefined) ? msg.body.meta.applicationUid : undefined;
+  try {
+    const token = await getToken(cfg);
+    const self = this;
+    const oihUid = (msg.body.meta !== undefined && msg.body.meta.oihUid !== undefined) ? msg.body.meta.oihUid : 'oihUid not set yet';
+    let recordUid = (msg.body !== undefined && msg.body.meta.recordUid !== undefined) ? msg.body.meta.recordUid : undefined;
+    const applicationUid = (msg.body.meta !== undefined && msg.body.meta.applicationUid !== undefined) ? msg.body.meta.applicationUid : undefined;
 
-  async function emitData() {
     /** Create an OIH meta object which is required
     * to make the Hub and Spoke architecture work properly
     */
@@ -77,9 +76,6 @@ async function processAction(msg, cfg) {
       }
     }
 
-    console.log(`About to upsert ${type}: `);
-    console.log(JSON.stringify(contactObject));
-
     // Upsert the object depending on 'objectExists' property
     const reply = await upsertObject(contactObject, token, objectExists, type, recordUid, cfg.devMode);
 
@@ -99,33 +95,11 @@ async function processAction(msg, cfg) {
     }
 
     self.emit('data', newMessage(newElement));
-  }
-
-  /**
-   * This method will be called from OIH platform if an error occured
-   *
-   * @param e - object containg the error
-   */
-  function emitError(e) {
+  } catch (e) {
     console.error('ERROR: ', e);
     console.log('Oops! Error occurred');
     self.emit('error', e);
   }
-
-  /**
-   * This method will be called from OIH platform
-   * when the execution is finished successfully
-   *
-   */
-  function emitEnd() {
-    console.log('Finished execution');
-    self.emit('end');
-  }
-
-  Q()
-    .then(emitData)
-    .fail(emitError)
-    .done(emitEnd);
 }
 
 module.exports = {
