--- lib/schemas/getOrganizations.out.json
@@ -1,177 +1,0 @@
-{
-  "type": "object",
-  "properties": {
-    "uid": {
-      "type": "string",
-      "required": true,
-      "title": "Uid",
-      "note": "Organization's uid",
-      "placeholder": "3gbdq1jtsvxjot"
-    },
-    "name": {
-      "type": "string",
-      "required": true,
-      "title": "Name",
-      "note": "Organization's name",
-      "placeholder": "Company Ltd."
-    },
-    "addresses": {
-      "type": "array",
-      "required": true,
-      "title": "Addresses",
-      "note": "Organizations's addresses",
-      "items": {
-        "type": "object",
-        "properties": {
-          "uid": {
-            "type": "string",
-            "example": "3gbbu1jmkewc87"
-          },
-          "street": {
-            "type": "string",
-            "example": "Wendenstr"
-          },
-          "streetNumber": {
-            "type": "string",
-            "example": "120"
-          },
-          "unit": {
-            "type": "string",
-            "example": "Hammerbrook"
-          },
-          "zipcode": {
-            "type": "string",
-            "example": "20537"
-          },
-          "city": {
-            "type": "string",
-            "example": "Hamburg"
-          },
-          "district": {
-            "type": "string",
-            "example": "HH-Hamburg"
-          },
-          "region": {
-            "type": "string",
-            "example": "Hamburg"
-          },
-          "country": {
-            "type": "string",
-            "example": "Germany"
-          },
-          "countryCode": {
-            "type": "string",
-            "example": "0049"
-          },
-          "primaryContact": {
-            "type": "string",
-            "example": "true"
-          },
-          "description": {
-            "type": "string",
-            "example": "Business Address"
-          }
-        }
-      }
-    },
-    "categories": {
-      "type": "array",
-      "required": true,
-      "title": "Categories",
-      "note": "Organizations's categories",
-      "items": {
-        "type": "object",
-        "properties": {
-          "uid": {
-            "type": "string",
-            "example": "yj5rl1jt7j9ce2"
-          },
-          "label": {
-            "type": "string",
-            "example": "Customers"
-          }
-        }
-      }
-    },
-    "relations": {
-      "type": "array",
-      "required": true,
-      "title": "Relations",
-      "note": "Organizations's relations",
-      "items": {
-        "type": "object",
-        "properties": {
-          "uid": {
-            "type": "string",
-            "example": "yj5pv1jtva0k0t"
-          },
-          "type": {
-            "type": "string",
-            "example": "OrganizationToOrganization"
-          },
-          "typeRef": {
-            "type": "string",
-            "example": "yj5rl1jt01l18b"
-          },
-          "label": {
-            "type": "string",
-            "example": "Partner"
-          },
-          "uids": {
-            "type": "array",
-            "items": {}
-          },
-          "partner": {
-            "type": "object",
-            "properties": {
-              "kind": {
-                "type": "string",
-                "example": "Organization"
-              },
-              "name": {
-                "type": "string",
-                "example": "Partner Company Ltd."
-              },
-              "uid": {
-                "type": "string",
-                "example": "3gbdq1jtss29zz"
-              },
-              "addresses": {
-                "type": "array",
-                "items": {}
-              }
-            }
-          }
-        }
-      }
-    },
-    "contactData": {
-      "type": "array",
-      "items": {
-        "type": "object",
-        "properties": {
-          "uid": {
-            "type": "string",
-            "example": "3gbdq1jtss6kc4"
-          },
-          "description": {
-            "type": "string",
-            "example": "Support email"
-          },
-          "value": {
-            "type": "string",
-            "example": "support@partnercompany.com"
-          },
-          "type": {
-            "type": "string",
-            "example": "email"
-          },
-          "categories": {
-            "type": "array",
-            "items": {}
-          }
-        }
-      }
-    }
-  }
-}
