--- lib/schemas/getPersons.out.json
@@ -1,282 +1,0 @@
-{
-  "type": "object",
-  "properties": {
-    "meta": {
-      "type": "object",
-      "properties": {
-        "username": {
-          "type": "string",
-          "example": "jdoe@mail.com"
-        },
-        "tenant": {
-          "type": "string",
-          "example": "5c7f8c5ea3c613001042f619"
-        },
-        "user": {
-          "type": "string",
-          "example": "5c7f8c5da3c613001042f618"
-        },
-        "role": {
-          "type": "string",
-          "example": "USER"
-        }
-      }
-    },
-    "uid": {
-      "type": "string",
-      "required": true,
-      "title": "Uid",
-      "note": "Person's uid",
-      "placeholder": "3gbdq1jtsvxjot"
-    },
-    "title": {
-      "type": "string",
-      "required": true,
-      "title": "Title",
-      "note": "Person's title",
-      "placeholder": "Prof."
-    },
-    "salutation": {
-      "type": "string",
-      "required": true,
-      "title": "Salutation",
-      "note": "Person's salutation",
-      "placeholder": "Herr"
-    },
-    "firstName": {
-      "type": "string",
-      "required": true,
-      "title": "First name",
-      "note": "Person's last name",
-      "placeholder": "Doe"
-    },
-    "lastName": {
-      "type": "string",
-      "required": true,
-      "title": "Name",
-      "note": "Person's last name",
-      "placeholder": "Doe"
-    },
-    "middleName": {
-      "type": "string",
-      "required": true,
-      "title": "Name",
-      "note": "Person's middle name",
-      "placeholder": "Anthony"
-    },
-    "gender": {
-      "type": "string",
-      "required": true,
-      "title": "Gender",
-      "note": "Gender",
-      "placeholder": "Male"
-    },
-    "birthday": {
-      "type": "string",
-      "required": true,
-      "title": "Birthday",
-      "note": "Person's date of birth",
-      "placeholder": "09.11.1980"
-    },
-    "notes": {
-      "type": "string",
-      "required": true,
-      "title": "Notes",
-      "note": "Person's notes",
-      "placeholder": "Some notes"
-    },
-    "displayName": {
-      "type": "string",
-      "required": true,
-      "title": "Display name",
-      "note": "Display name",
-      "placeholder": "John"
-    },
-    "language": {
-      "type": "string",
-      "required": true,
-      "title": "Language",
-      "note": "Preferred language",
-      "placeholder": "English"
-    },
-    "nickname": {
-      "type": "string",
-      "required": true,
-      "title": "Nickname",
-      "note": "Display nickname",
-      "placeholder": "Johny"
-    },
-    "jobTitle": {
-      "type": "string",
-      "required": true,
-      "title": "Job title",
-      "note": "Job title",
-      "placeholder": "Manager"
-    },
-    "photo": {
-      "type": "string",
-      "required": true,
-      "title": "Photo",
-      "note": "Person's photo",
-      "placeholder": "www.images.com/john.png"
-    },
-    "addresses": {
-      "type": "array",
-      "required": true,
-      "title": "Addresses",
-      "note": "Person's addresses",
-      "items": {
-        "type": "object",
-        "properties": {
-          "uid": {
-            "type": "string",
-            "example": "3gbbu1jmkewc87"
-          },
-          "street": {
-            "type": "string",
-            "example": "Wendenstr"
-          },
-          "streetNumber": {
-            "type": "string",
-            "example": "120"
-          },
-          "unit": {
-            "type": "string",
-            "example": "Hammerbrook"
-          },
-          "zipcode": {
-            "type": "string",
-            "example": "20537"
-          },
-          "city": {
-            "type": "string",
-            "example": "Hamburg"
-          },
-          "district": {
-            "type": "string",
-            "example": "HH-Hamburg"
-          },
-          "region": {
-            "type": "string",
-            "example": "Hamburg"
-          },
-          "country": {
-            "type": "string",
-            "example": "Germany"
-          },
-          "countryCode": {
-            "type": "string",
-            "example": "0049"
-          },
-          "primaryContact": {
-            "type": "string",
-            "example": "true"
-          },
-          "description": {
-            "type": "string",
-            "example": "Private Address"
-          }
-        }
-      }
-    },
-    "categories": {
-      "type": "array",
-      "required": true,
-      "title": "Categories",
-      "note": "Person's categories",
-      "items": {
-        "type": "object",
-        "properties": {
-          "uid": {
-            "type": "string",
-            "example": "yj5rl1jt7j9ce2"
-          },
-          "label": {
-            "type": "string",
-            "example": "Private"
-          }
-        }
-      }
-    },
-    "relations": {
-      "type": "array",
-      "required": true,
-      "title": "Relations",
-      "note": "Person's relations",
-      "items": {
-        "type": "object",
-        "properties": {
-          "uid": {
-            "type": "string",
-            "example": "yj5pv1jtva0k0t"
-          },
-          "type": {
-            "type": "string",
-            "example": "OrganizationToPerson"
-          },
-          "typeRef": {
-            "type": "string",
-            "example": "yj5rl1jt01l18b"
-          },
-          "label": {
-            "type": "string",
-            "example": "CEO"
-          },
-          "uids": {
-            "type": "array",
-            "items": {}
-          },
-          "partner": {
-            "type": "object",
-            "properties": {
-              "kind": {
-                "type": "string",
-                "example": "Organization"
-              },
-              "name": {
-                "type": "string",
-                "example": "Company Ltd."
-              },
-              "uid": {
-                "type": "string",
-                "example": "3gbdq1jtss29zz"
-              },
-              "addresses": {
-                "type": "array",
-                "items": {}
-              }
-            }
-          }
-        }
-      }
-    },
-    "contactData": {
-      "type": "array",
-      "items": {
-        "type": "object",
-        "properties": {
-          "uid": {
-            "type": "string",
-            "example": "3gbdq1jtss6kc4"
-          },
-          "description": {
-            "type": "string",
-            "example": "Information email"
-          },
-          "value": {
-            "type": "string",
-            "example": "info@company.com"
-          },
-          "type": {
-            "type": "string",
-            "example": "email"
-          },
-          "categories": {
-            "type": "array",
-            "items": {}
-          }
-        }
-      }
-    }
-  }
-}
