--- lib/schemas/upsertOrganization.in.json
@@ -1,26 +1,0 @@
-{
-  "type": "object",
-  "properties": {
-    "uid": {
-      "type": "string",
-      "required": false,
-      "title": "Uid",
-      "note": "Person's uid",
-      "placeholder": "25mop1jxq2ss3x"
-    },
-    "name": {
-      "type": "string",
-      "required": true,
-      "title": "Organization's name",
-      "note": "Organization's name",
-      "placeholder": "Company Ltd."
-    },
-    "logo": {
-      "type": "string",
-      "required": false,
-      "title": "Logo",
-      "note": "Organization's logo",
-      "placeholder": "www.company.com/logo.png"
-    }
-  }
-}
