--- lib/schemas/upsertPerson.in.json
@@ -1,82 +1,0 @@
-{
-  "type": "object",
-  "properties": {
-    "uid": {
-      "type": "string",
-      "required": false,
-      "title": "Uid",
-      "note": "Person's uid",
-      "placeholder": "25mop1jxq2ss3x"
-    },
-    "firstName": {
-      "type": "string",
-      "required": true,
-      "title": "First name",
-      "note": "Person's last name",
-      "placeholder": "Doe"
-    },
-    "lastName": {
-      "type": "string",
-      "required": true,
-      "title": "Last Name",
-      "note": "Person's last name",
-      "placeholder": "Doe"
-    },
-    "middleName": {
-      "type": "string",
-      "required": false,
-      "title": "Middle Name",
-      "note": "Person's middle name",
-      "placeholder": "Anthony"
-    },
-    "email": {
-      "type": "string",
-      "required": false,
-      "title": "Email",
-      "note": "Person's email",
-      "placeholder": "johndoe@workmail.com"
-    },
-    "title": {
-      "type": "string",
-      "required": false,
-      "title": "Title",
-      "note": "Person's title",
-      "placeholder": "Prof."
-    },
-    "salutation": {
-      "type": "string",
-      "required": false,
-      "title": "Salutation",
-      "note": "Person's salutation",
-      "placeholder": "Herr"
-    },
-    "gender": {
-      "type": "string",
-      "required": false,
-      "title": "Gender",
-      "note": "Person's gender",
-      "placeholder": "Male"
-    },
-    "position": {
-      "type": "string",
-      "required": false,
-      "title": "Position / Business card information",
-      "note": "Person's position in the company",
-      "placeholder": "Manager"
-    },
-    "birthday": {
-      "type": "string",
-      "required": false,
-      "title": "Birthday",
-      "note": "Person's date of birth",
-      "placeholder": "09.11.1980"
-    },
-    "nickname": {
-      "type": "string",
-      "required": false,
-      "title": "Nickname",
-      "note": "Person's nickname",
-      "placeholder": "Johny"
-    }
-  }
-}
