--- lib/schemas/upsertPerson.out.json
@@ -1,180 +1,0 @@
-{
-  "type": "object",
-  "properties": {
-    "rowid": {
-      "type": "string",
-      "required": true,
-      "title": "Rowid",
-      "note": "Person's rowid in Snazzy Contacts",
-      "placeholder": "199997"
-    },
-    "firstname": {
-      "type": "string",
-      "required": true,
-      "title": "First name",
-      "note": "Person's last name",
-      "placeholder": "Doe"
-    },
-    "name": {
-      "type": "string",
-      "required": true,
-      "title": "Name",
-      "note": "Person's last name",
-      "placeholder": "Doe"
-    },
-    "email": {
-      "type": "string",
-      "required": false,
-      "title": "Email",
-      "note": "Person's email",
-      "placeholder": "johndoe@workmail.com"
-    },
-    "title": {
-      "type": "string",
-      "required": false,
-      "title": "Title",
-      "note": "Person's title",
-      "placeholder": "Prof."
-    },
-    "salutation": {
-      "type": "string",
-      "required": false,
-      "title": "Salutation",
-      "note": "Person's salutation",
-      "placeholder": "Herr"
-    },
-    "position": {
-      "type": "string",
-      "required": false,
-      "title": "Position / Business card information",
-      "note": "Person's position in the company",
-      "placeholder": "Manager"
-    },
-    "birthday": {
-      "type": "string",
-      "required": false,
-      "title": "Birthday",
-      "note": "Person's date of birth",
-      "placeholder": "09.11.1980"
-    },
-    "private_street": {
-      "type": "string",
-      "required": false,
-      "title": "Street",
-      "note": "Person's address street",
-      "placeholder": "Main Str. 120"
-    },
-    "private_zip_code": {
-      "type": "string",
-      "required": false,
-      "title": "Zip code",
-      "note": "Person's address zip code",
-      "placeholder": "28507"
-    },
-    "private_town": {
-      "type": "string",
-      "required": false,
-      "title": "Town",
-      "note": "Person's town",
-      "placeholder": "Hamburg"
-    },
-    "private_country": {
-      "type": "string",
-      "required": false,
-      "title": "Country",
-      "note": "Person's country",
-      "placeholder": "Germany"
-    },
-    "fax": {
-      "type": "string",
-      "required": false,
-      "title": "Fax",
-      "note": "Person's fax",
-      "placeholder": "0407130199"
-    },
-    "phone": {
-      "type": "string",
-      "required": false,
-      "title": "Phone",
-      "note": "Person's phone",
-      "placeholder": "0407130170"
-    },
-    "mobile_phone": {
-      "type": "string",
-      "required": false,
-      "title": "Mobile phone",
-      "note": "Person's mobile phone",
-      "placeholder": "017631046821"
-    },
-    "private_mobile_phone": {
-      "type": "string",
-      "required": false,
-      "title": "Private mobile phone",
-      "note": "Person's private mobile phone",
-      "placeholder": "017631926168"
-    },
-    "private_phone": {
-      "type": "string",
-      "required": false,
-      "title": "Private phone",
-      "note": "Person's private phone",
-      "placeholder": "017131996877"
-    },
-    "facebook_url": {
-      "type": "string",
-      "required": false,
-      "title": "Facebook",
-      "note": "Person's facebook account",
-      "placeholder": "www.facebook.com/johndoe"
-    },
-    "linked_in_url": {
-      "type": "string",
-      "required": false,
-      "title": "Linkedin",
-      "note": "Person's linkedin account",
-      "placeholder": "www.linkedin.com/johndoe"
-    },
-    "twitter_url": {
-      "type": "string",
-      "required": false,
-      "title": "Twitter",
-      "note": "Person's twitter account",
-      "placeholder": "www.twitter.com/johndoe"
-    },
-    "googleplus_url": {
-      "type": "string",
-      "required": false,
-      "title": "Google plus",
-      "note": "Person's google plus account",
-      "placeholder": "www.plus.google.com/johndoe"
-    },
-    "youtube_url": {
-      "type": "string",
-      "required": false,
-      "title": "Youtube",
-      "note": "Person's youtube account",
-      "placeholder": "johndoe@mail.com"
-    },
-    "url": {
-      "type": "string",
-      "required": false,
-      "title": "Website",
-      "note": "Person's website",
-      "placeholder": "www.johndoe.com"
-    },
-    "skype": {
-      "type": "string",
-      "required": false,
-      "title": "Skype",
-      "note": "Person's skype account",
-      "placeholder": "john.doe"
-    },
-    "for_rowid": {
-      "type": "string",
-      "required": false,
-      "title": "Organization ID",
-      "note": "Organization's id where the person is part of",
-      "placeholder": "200305"
-    }
-  }
-}
