+++ lib/triggers/getOrganizationsPolling.js
@@ -16,9 +16,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-const Q = require('q');
-const { newMessage } = require('../helpers');
-const { getEntries } = require('./../utils/helpers');
+const { getEntries, newMessage } = require('./../utils/helpers');
 const { getToken } = require('./../utils/authentication');
 
 /**
@@ -29,16 +27,16 @@ const { getToken } = require('./../utils/authentication');
  * @param snapshot - saves the current state of integration step for the future reference
  */
 async function processTrigger(msg, cfg, snapshot = {}) {
+  try {
   // Authenticate and get the token from Snazzy Contacts
-  const { applicationUid, domainId, schema } = cfg;
-  // const token = cfg.API_KEY;
-  const token = await getToken(cfg);
-  const self = this;
+    const { applicationUid, domainId, schema } = cfg;
+    // const token = cfg.API_KEY;
+    const token = await getToken(cfg);
+    const self = this;
 
-  // Set the snapshot if it is not provided
-  snapshot.lastUpdated = snapshot.lastUpdated || (new Date(0)).getTime();
+    // Set the snapshot if it is not provided
+    snapshot.lastUpdated = snapshot.lastUpdated || (new Date(0)).getTime();
 
-  async function emitData() {
     /** Create an OIH meta object which is required
     * to make the Hub and Spoke architecture work properly
     */
@@ -47,11 +45,6 @@ async function processTrigger(msg, cfg, snapshot = {}) {
       schema: (schema !== undefined && schema !== null) ? schema : undefined,
       domainId: (domainId !== undefined && domainId !== null) ? domainId : undefined,
     };
-
-    // Get the total amount of fetched objects
-    let count;
-    const getCount = await getEntries(token, snapshot, count, 'organization', cfg.devMode);
-    count = getCount.count; // eslint-disable-line
 
     const organizations = await getEntries(token, snapshot, count, 'organization', cfg.devMode);
 
@@ -86,35 +79,12 @@ async function processTrigger(msg, cfg, snapshot = {}) {
     } else {
       self.emit('snapshot', snapshot);
     }
-  }
-
-  /**
-   * This method will be called from OIH platform if an error occured
-   *
-   * @param e - object containg the error
-   */
-  function emitError(e) {
+  } catch (e) {
     console.log(`ERROR: ${e}`);
     self.emit('error', e);
   }
-
-  /**
-   * This method will be called from OIH platform
-   * when the execution is finished successfully
-   *
-   */
-  function emitEnd() {
-    console.log('Finished execution');
-    self.emit('end');
-  }
-
-  Q()
-    .then(emitData)
-    .fail(emitError)
-    .done(emitEnd);
 }
 
 module.exports = {
   process: processTrigger,
-  getEntries,
 };
