+++ lib/utils/helpers.js
@@ -263,6 +263,27 @@ async function getEntries(token, snapshot, count, type, devMode) {
   }
 }
 
+
+const uuid = require('uuid');
+
+function newMessage(body) {
+  const msg = {
+    id: uuid.v4(),
+    attachments: {},
+    body,
+    headers: {},
+    metadata: {},
+  };
+
+  return msg;
+}
+
+
 module.exports = {
-  getEntries, upsertObject, fetchUidFromReference, populateRelations,
+  newMessage,
+};
+
+
+module.exports = {
+  getEntries, upsertObject, fetchUidFromReference, populateRelations, newMessage,
 };
